refactor : 폴더 구조를 도메인 중심으로 변경한다
