+++ src/test/java/subway/station/StationTestFixture.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.station;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
@@ -9,7 +9,7 @@ import org.springframework.http.MediaType;
 import java.util.HashMap;
 import java.util.Map;
 
-public class SubwayTestFixture {
+public class StationTestFixture {
     public static ExtractableResponse<Response> 지하철_역_생성_요청(String name) {
         Map<String, String> params = new HashMap<>();
         params.put("name", name);
